nv_core/validator_sample/requirements.txt
omniverse-asset-validator[usd,numpy]
omniverse-usd-profiles
usd-core>=22.11